% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_javascript.R
\name{add_javascript}
\alias{add_javascript}
\title{Add necessary JavaScript code to enable interactivity}
\usage{
add_javascript(in_html_doc = TRUE, filepath = NULL)
}
\arguments{
\item{in_html_doc}{A single value, TRUE or FALSE, indicating whether the JavaScript should be included in the HTML document or if instead a separate '.js' file should be created.}

\item{filepath}{If `in_html_doc = FALSE`, the filepath of the '.js' JavaScript file to be created.}
}
\value{
Either outputs raw HTML/JavaScript code that can live inside an HTML file, or creates a '.js' file containing the JavaScript code.
}
\description{
Add necessary JavaScript code to enable interactivity
}
\examples{
\dontrun{
# RMarkdown code

```{r add_js, results='asis'}
library(selr)
add_javascript()
```

```{r add_selector, results='asis'}
add_selector(
sel_id = "Language",
options = list("Option_1" = "English",
               "Option_2" = "Hebrew",
               "Option_3" = "Pirate")
)
```

```{r, results='asis'}
sel_text(
sel_id = "Language",
options = list("Option_1" = "Hello",
               "Option_2" = "Shalom",
               "Option_3" = "Ahoy")
)
```
}

}
