% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selr-data.R
\docType{data}
\name{bitterness}
\alias{bitterness}
\title{White Wine Bitterness}
\format{
The \code{bitterness} data frame has 72 rows and 3 columns.
}
\usage{
bitterness
}
\description{
White Wine Bitterness
}
\details{
Data of a study on the factors influencing white wine bitterness. Bitterness
is measured by 1~5 level from "least bitter" to "most bitter". This data frame
has already been reformatted for SELR package usage.



\describe{
\item{temperature}{The temperature controlled during crushing the grapes ('warm' or 'cold')}
\item{contact}{The contact controlled during crushing the grapes ('yes' or 'no')}
\item{bitterness}{The bitterness of the white wine (1~5)}
}
}
\examples{

data(bitterness)
head(bitterness)

mod1 <- selr(bitterness~temperature*contact,data=bitterness,modelType="cumulative")
summary(mod1)

mod2 <- selr(bitterness~temperature+contact,data=bitterness,modelType="cumulative",parallel = FALSE)
summary(mod2)

mod3 <- selr(bitterness~temperature+contact,data=bitterness,modelType="continuation",elr=TRUE)
summary(mod3)

anova(mod1,test="score")

QIC(mod1)
QIC(mod1,mod2)

}
\references{
LRandall, J. H. (1989). The Analysis of Sensory Data by Generalized Linear Model.
Biometrical Journal, 31(7), 781–793. https://doi.org/10.1002/bimj.4710310703
}
\keyword{datasets}
