% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selr-data.R
\docType{data}
\name{mntlhlth}
\alias{mntlhlth}
\title{Mental Health Status}
\format{
The \code{mntlhlth} data frame has 1660 rows and 2 columns.
}
\usage{
mntlhlth
}
\description{
Mental Health Status
}
\details{
Data of a study on the relationship between an individual’s mental health status
and the socioeconomic status of his or her parents. 'socioeconomic' has 0 to 5 levels,
from 'lowest' to 'highest' and 'mntlhlth' has four levels: 'well', 'mild', 'moderate', 'impaired'



\describe{
\item{socioeconomic}{The socioeconomic status of the individual's parents (0~5)}
\item{mntlhlth}{an individual’s mental health status ('well', 'mild', 'moderate', 'impaired')}
}
}
\examples{

data(mntlhlth)
head(mntlhlth)

mod1 <- selr(mntlhlth~socioeconomic,data=mntlhlth,modelType="cumulative",parallel=FALSE)
summary(mod1)

mod2 <- selr(mntlhlth~socioeconomic,data=mntlhlth,modelType="cumulative")
summary(mod2)

anova(mod1,mod2,test="score")
anova(mod1,test="score")

QIC(mod1)
QIC(mod1,mod2)

mod3 <- selr(mntlhlth~socioeconomic,data=mntlhlth,modelType="cumulative",parallel=FALSE,elr=TRUE)
summary(mod3)

}
\references{
Fu, L., & Simpson, D. G. (2002).
Conditional risk models for ordinal response data:
simultaneous logistic regression analysis and generalized score tests.
Journal of Statistical Planning and Inference, 108(1–2), 201–217.
https://doi.org/10.1016/s0378-3758(02)00279-3
}
\keyword{datasets}
