% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selr.R
\name{predict.selr}
\alias{predict.selr}
\title{Predictions for SELR}
\usage{
\method{predict}{selr}(object, newdata)
}
\arguments{
\item{object}{a \code{selr} object}

\item{newdata}{a data set containing the explanatory variables}

\item{type}{a logical variable, available options are \code{'link'}, \code{'response'}}

\item{marginal}{a logical variable, indicating whether marginal probability
should be calculated. If not, conditional risk probbaility would be returned.}
}
\value{
\code{predict.selr} returns a data set containing the predicted values.
}
\description{
Predict function for \code{selr} model.
}
\examples{

data(mntlhlth)
head(mntlhlth)

set.seed(123)
mntlhlth$mntlhlth <- factor(mntlhlth$mntlhlth, levels = c("well", "mild", "moderate", "impaired"))
mntlhlth$socioeconomic = as.numeric(mntlhlth$socioeconomic)
itrain=sample(seq(nrow(mntlhlth)),floor(0.8*nrow(mntlhlth)),replace=FALSE)
train=mntlhlth[itrain,]
test=mntlhlth[-itrain,]
ord <- match(train$mntlhlth, list(1,2,3,4))
train <- train[order(ord),]
train <- train[order(train$socioeconomic),]
ord <- match(test$mntlhlth, list(1,2,3,4))
test <- test[order(ord),]
test <- test[order(test$socioeconomic),]
head(train)
head(test)

mod1 <- selr(mntlhlth~socioeconomic,data=train,modelType="continuation")
summary(mod1)

data<-data.frame(test[,-2])
colnames(data)<-c("socioeconomic")
pred <- predict(object=mod1,newdata=data)
print(pred)


}
\seealso{
\code{\link{predict.glm}}
}
\author{
Jiyang Xu \email{jiyangx3@illinois.edu},
        Douglas Simpson \email{dgs@illinois.edu}
}
\keyword{predict}
