% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selr-data.R
\docType{data}
\name{ptb}
\alias{ptb}
\title{Preterm Birth}
\format{
The \code{ptb} data frame has 424 rows and 17 columns.
}
\usage{
ptb
}
\description{
Preterm Birth
}
\details{
Data of a study on the factors influencing preterm birth. The response
are nominal factors: "FTB", "sPTB", "mPTB", where "FTB" for Full-Term Birth,
"sPTB" for Spontaneous Preterm Birth, "mPTB" for Moderate Preterm Birth.
}
\examples{

data(ptb)
head(ptb)

}
\keyword{datasets}
