% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selr.R
\name{sa_RPS}
\alias{sa_RPS}
\title{Squared Absolute RPS}
\usage{
sa_RPS(p, y)
}
\arguments{
\item{p}{a matrix of predicted probabilities, where each column corresponds
to one level}

\item{y}{a matrix of one-hot representation for observed outcomes}
}
\value{
\code{sa_RPS} returns a vector of squared absolute rps for each
observation
}
\description{
A modified version of ranked probability score for ordinal
classification
}
\references{
Galdran, A. (2023). Performance Metrics for Probabilistic Ordinal Classifiers.
In: Greenspan, H., et al. Medical Image Computing and Computer Assisted
Intervention – MICCAI 2023. MICCAI 2023. Lecture Notes in Computer Science,
vol 14222. Springer, Cham. https://doi.org/10.1007/978-3-031-43898-1_35
}
\seealso{
\code{\link[verification]{rps}}
}
\author{
Jiyang Xu \email{jiyangx3@illinois.edu},
        Douglas Simpson \email{dgs@illinois.edu}
}
