% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sel_image.R
\name{sel_image}
\alias{sel_image}
\title{Create a dynamic image object}
\usage{
sel_image(sel_id = "ts_1", options = c(Option_1 = "Images/Iris.png",
  Option_2 = "Images/mtcars.png", Option_3 = "Images/US Arrests.png"),
  embed = TRUE)
}
\arguments{
\item{sel_id}{The id of the selector used to choose options}

\item{options}{A named list of filepaths, with the number of options and names of options matching those used in the selector object.}

\item{embed}{A logical vector for whether images should be embedded directly into the HTML code as base64 strings. If FALSE, image files are referred to via links in the HTML code.}
}
\value{
Raw HTML representing the image object
}
\description{
Create a dynamic image object
}
\examples{

```{r add_js, results='asis'}
library(selr)
add_javascript()
```

```{r add_selector, results='asis'}
add_selector(
sel_id = "Dataset",
options = list("Option_1" = "iris",
               "Option_2" = "mtcars",
               "Option_3" = "US Arrests")
)
```

```{r, results='asis'}
sel_text(
sel_id = "Dataset",
options = list("Option_1" = "Plots/Iris.png",
               "Option_2" = "Plots/mtcars.png",
               "Option_3" = "Plots/US Arrests.png")
)
```

}
