% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_selector.R
\name{add_selector}
\alias{add_selector}
\title{Title}
\usage{
add_selector(sel_id = "ts_1", options = c(Option_1 = "Option #1",
  Option_2 = "Option #2", Option_3 = "Option #3"),
  label = "Select an option:")
}
\arguments{
\item{sel_id}{A unique ID for the selector which can be referenced in other places.}

\item{options}{A named character vector. The elements of the character vector are the text visible in the select object. The names serve as option IDs which can be referenced in other places.}

\item{label}{The text to display along with the selector.}
}
\value{
Lines of HTML text which function as a selector.
}
\description{
Title
}
